break tools/o/index.js into own files
