+++ middleware/tools/o/messages.js
@@ -1,0 +1,20 @@
+/***********************************************************
+  messages
+***********************************************************/
+
+const { message } = require ('../')
+
+/***************************************
+  definition
+***************************************/
+
+const messages = {
+  'hello' :
+    () => (ri, ro) => (message (
+      `hello! i'm ~${ri.originalUrl}.`
+    )),
+}
+
+/**************************************/
+
+module.exports = messages
